Changes 'window already there' action in WPXA_Move Removed functionality where window jumps to next monitor when at edge. Or shrink if no next monitor available. Now it just resize height instead to give an additional height option.
Note that the heights are written directly in the function and can therefor currently not be passed as argument.
It is simply based on whether the vertical offset passed as argument is +1, 0 or -1.

E.g.
If SideY (vertical offset) is -1 (top) heightFactor becomes 70 percent for a window that is allready in its requested location.